hubble: Update uint size in flow proto
Sizes of some fields in flow.proto is unnecessary
large (uint64), whereas for such fields
(e.g. endpoint ID or identity) a smaller type
is more appropriate.

This PR is a clean-up to update the scalar
value types of these fields. Includes some
changes to related Hubble packages (parsers
and filters) and minor test adjustments.

Fixes: cilium/hubble#158

+++ pkg/hubble/filters/filters_test.go
@@ -96,7 +96,7 @@ func (t *testFilterFalse) OnBuildFilter(_ context.Context, ff *pb.FlowFilter) ([
 
 func TestOnBuildFilter(t *testing.T) {
 	fl, err := BuildFilterList(context.Background(),
-		[]*pb.FlowFilter{{SourceIdentity: []uint64{1, 2, 3}}}, // true
+		[]*pb.FlowFilter{{SourceIdentity: []uint32{1, 2, 3}}}, // true
 		[]OnBuildFilter{&testFilterTrue{}})                    // true
 	assert.NoError(t, err)
 	assert.Equal(t, true, fl.MatchAll(&v1.Event{Event: &pb.Flow{
@@ -104,7 +104,7 @@ func TestOnBuildFilter(t *testing.T) {
 	}}))
 
 	fl, err = BuildFilterList(context.Background(),
-		[]*pb.FlowFilter{{SourceIdentity: []uint64{1, 2, 3}}}, // true
+		[]*pb.FlowFilter{{SourceIdentity: []uint32{1, 2, 3}}}, // true
 		[]OnBuildFilter{&testFilterFalse{}})                   // false
 	assert.NoError(t, err)
 	assert.Equal(t, false, fl.MatchAll(&v1.Event{Event: &pb.Flow{
@@ -112,7 +112,7 @@ func TestOnBuildFilter(t *testing.T) {
 	}}))
 
 	fl, err = BuildFilterList(context.Background(),
-		[]*pb.FlowFilter{{SourceIdentity: []uint64{1, 2, 3}}}, // true
+		[]*pb.FlowFilter{{SourceIdentity: []uint32{1, 2, 3}}}, // true
 		[]OnBuildFilter{
 			&testFilterFalse{}, // false
 			&testFilterTrue{}}) // true
